requirements.txt
Flask==1.1.2
numpy==1.17.4
tensorflow==2.2.0
Pillow==6.2.2
gunicorn==20.0.4
